ml-predicter/requirements-loose.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
asyncpg
redis
pandas
numpy
scikit-learn
statsmodels
scipy
requests
python-dotenv
